requirements.txt
et-xmlfile==1.1.0
mysql-connector-python==8.0.33
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
XlsxWriter==3.1.0
